Make visual studio happy about a conversion from double to int.

+++ Source/lib/nd_hacdStructs.h
@@ -68,7 +68,7 @@ struct HACDDecoder: public HACD::ICallback
 	virtual void operator()( char const *aMsg, double aProgress, double aConcavity, size_t aVertices)
 	{
 		if( mCallback )
-			(*mCallback)(aMsg, aProgress, aVertices );
+			(*mCallback)(aMsg, static_cast<int>(aProgress), aVertices );
 	}
 
 };
